requirements.txt
git+https://github.com/sympy/sympy.git@master
scipy
wheel
maturin
SQLAlchemy
psycopg2-binary
python-dotenv